Fixed arduino code (machine readable), converted message from bytes to int, fixed typos
The arduino code was still setup to spit out human readable messages to the serial port.  Changed it back to plain binary data.  
Once the message is read in the python script, converted the bytes to int.  It is harder to deal with bytes.
Fixed a typo in the fields dictionary (the from byte needs to be bigger than the start byte... duh).
Still need to find the e-shower aperature voltage bytes?
Need to find a way to do formating on the different numbers (ints, floats, scientific notation, strings)